gemini-agent/pyproject.toml
[project]
name = "gemini-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.36.0",
    "python-dotenv>=1.1.1",
]
